requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
docutils==0.21.2
Flask==3.1.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
Pygments==2.18.0
PyMySQL==1.1.1
pypiwin32==223
pywin32==308
requests==2.32.3
SQLAlchemy==2.0.36
tabulate==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
